add transactions stream isolation
